backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
huggingface-hub==0.19.4
PyMuPDF==1.23.8
pdfplumber==0.10.3
python-docx==0.8.11
pillow==10.1.0
python-dotenv==1.0.0
pydantic==2.5.0
cors==1.0.1
pytesseract==0.3.10
requests==2.31.0
transformers==4.36.2
torch==2.1.2
